+++ packages/@tinacms/cli/src/next/commands/dev-command/index.ts
@@ -125,7 +125,8 @@ export class DevCommand extends BaseCommand {
 
           if (configManager.hasSeparateContentRoot()) {
             const rootPath = await configManager.getTinaFolderPath(
-              configManager.contentRootPath
+              configManager.contentRootPath,
+              { isContentRoot: true }
             );
             const filePath = path.join(rootPath, tinaLockFilename);
             await fs.ensureFile(filePath);
@@ -325,6 +326,12 @@ export class DevCommand extends BaseCommand {
         // },
       ],
     });
+    if (configManager?.config?.telemetry === 'anonymous') {
+      logger.info(
+        `\n📊 Note: TinaCMS now collects anonymous telemetry regarding usage. More information on TinaCMS Telemetry: https://tina.io/telemetry\n`
+      );
+    }
+
     await this.startSubCommand();
   }