+++ packages/tinacms-authjs/CHANGELOG.md
@@ -1,5 +1,35 @@
 # tinacms-authjs
 
+## 20.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - tinacms@3.6.1
+
+## 20.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`f23c2a6`](https://github.com/tinacms/tinacms/commit/f23c2a6742373271531bee46ca3cca8b208c3476), [`d782ca2`](https://github.com/tinacms/tinacms/commit/d782ca209ff98ee2143c7fd83f0b793563b6de60), [`56d533e`](https://github.com/tinacms/tinacms/commit/56d533e610a520ba66b3e58f3a0dc03487d5d5d7), [`60510bb`](https://github.com/tinacms/tinacms/commit/60510bbf8d1b78cad901722d4f9003665412d71e)]:
+  - tinacms@3.6.0
+  - @tinacms/schema-tools@2.7.0
+
+## 19.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`2738749`](https://github.com/tinacms/tinacms/commit/2738749bac08c02214fdf208f96bbcc4906ef442), [`dfb4ce8`](https://github.com/tinacms/tinacms/commit/dfb4ce8da7b82b638337109cc7859125e112a0a9)]:
+  - tinacms@3.5.1
+
+## 19.0.0
+
+### Patch Changes
+
+- Updated dependencies [[`da837f7`](https://github.com/tinacms/tinacms/commit/da837f7b17f0540cd47025afd2c45140706d595d), [`f90d47b`](https://github.com/tinacms/tinacms/commit/f90d47b746b000e65324da430aacc8fd623f5065), [`c806e41`](https://github.com/tinacms/tinacms/commit/c806e41232e0204a13539d6236358cb58cbc7692), [`89aee8a`](https://github.com/tinacms/tinacms/commit/89aee8ae3d6867e5793e4df7ff9c700dfd8a48c9)]:
+  - tinacms@3.5.0
+  - @tinacms/schema-tools@2.6.0
+
 ## 18.0.1
 
 ### Patch Changes